+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * `Licensee` enumeration with the description of the publisher of the ROM
+* Support for Zippy encoding format for fast compression
 
 ### Changed